\name{gdmExpData}
\alias{gdmExpData}
\docType{data}
\title{
Species and environmental data used in GDM examples
}
\description{
A dataset containing the species, site, and environmental data used to fit GDM in examples.
}
\format{
A data frame with 42,142 rows and 14 variables
}
\details{
-species: Species identifier.

-site: Site identification.

-awcA: plant-available water capacity of soil

-phTotal: soil pH

-sandA: percent sand content of soil

-shcA: saturated hydraulic conductivity of soil

-solumDepth: depth of solum layer (unknown units)

-bio5: Maximum Temperature of Warmest Month (C x 10)

-bio6: Minimum Temperature of Coldest Month (C x 10)

-bio15: Precipitation Seasonality (Coefficient of Variation)

-bio18: Precipitation of Warmest Quarter (mm)

-bio19: Precipitation of Coldest Quarter (mm)

-Lat: Latitude Coordinate of site

-Long: Longitude Coordinate of site
}
\source{
Fitzpatrick MC, Sanders NJ, Normand S, Svenning J-C, Ferrier S, Gove AD, Dunn RR (2013). Environmental and historical imprints on beta diversity: insights from variation in rates of species turnover along gradients. Proceedings of the Royal Society: Series B 280, art. 1768
}
\keyword{datasets}
